Split Droid adapter responsibilities
