+++ apps/server/src/provider/Layers/DroidAdapter.ts
@@ -1,280 +1,60 @@
 import { randomUUID } from "node:crypto";
 import {
-  AutonomyLevel,
   type AskUserRequestParams,
   type AskUserResult,
-  type Base64ImageSource,
-  type ContentBlock,
   createSession,
-  type CreateSessionOptions,
-  DroidInteractionMode,
-  DroidMessageType,
   type MessageOptions,
-  ReasoningEffort,
   resumeSession,
-  type ResumeSessionOptions,
   ToolConfirmationOutcome,
-  ToolConfirmationType,
-  type DroidMessage,
-  type DroidSession,
   type RequestPermissionRequestParams,
-  type TokenUsageUpdate,
 } from "@factory/droid-sdk";
 import {
   ApprovalRequestId,
-  EventId,
-  ProviderDriverKind,
   ProviderInstanceId,
-  RuntimeItemId,
-  RuntimeRequestId,
   ThreadId,
   TurnId,
-  type CanonicalRequestType,
   type DroidSettings,
-  type ProviderApprovalDecision,
   type ProviderRuntimeEvent,
   type ProviderSession,
-  type ProviderSessionStartInput,
-  type ProviderUserInputAnswers,
-  type RuntimeContentStreamKind,
-  type ToolLifecycleItemType,
-  type UserInputQuestion,
 } from "@t3tools/contracts";
 import { getModelSelectionStringOptionValue } from "@t3tools/shared/model";
 import * as Effect from "effect/Effect";
-import * as DateTime from "effect/DateTime";
 import * as FileSystem from "effect/FileSystem";
 import * as Queue from "effect/Queue";
 import * as Stream from "effect/Stream";
 
-import { resolveAttachmentPath } from "../../attachmentStore.ts";
 import { ServerConfig } from "../../config.ts";
 import {
-  type ProviderAdapterError,
   ProviderAdapterRequestError,
   ProviderAdapterSessionNotFoundError,
   ProviderAdapterValidationError,
 } from "../Errors.ts";
-import type { ProviderAdapterShape } from "../Services/ProviderAdapter.ts";
+import { resolveDroidImages } from "../droid/DroidAttachmentResolver.ts";
+import {
+  DROID_PROVIDER,
+  type DroidAdapterOptions,
+  type DroidAdapterShape,
+  type DroidContext,
+} from "../droid/DroidAdapterTypes.ts";
+import {
+  handleDroidMessage,
+  makeDroidEventBase,
+  nowIso,
+  updateDroidContextSession,
+} from "../droid/DroidRuntimeEvents.ts";
+import {
+  DroidInteractionMode,
+  normalizeAskUserQuestions,
+  permissionDetail,
+  toAskUserResult,
+  toAutonomyLevel,
+  toModelId,
+  toOutcome,
+  toReasoningEffort,
+  toRequestType,
+} from "../droid/DroidSdkMappings.ts";
 
-const PROVIDER = ProviderDriverKind.make("droid");
-
-interface PendingPermission {
-  readonly requestType: CanonicalRequestType;
-  readonly resolve: (decision: ToolConfirmationOutcome) => void;
-}
-
-interface PendingUserInput {
-  readonly questions: ReadonlyArray<UserInputQuestion>;
-  readonly droidQuestions: AskUserRequestParams["questions"];
-  readonly resolve: (result: AskUserResult) => void;
-}
-
-interface DroidContext {
-  session: ProviderSession;
-  readonly droid: DroidSession;
-  readonly pendingPermissions: Map<ApprovalRequestId, PendingPermission>;
-  readonly pendingUserInputs: Map<ApprovalRequestId, PendingUserInput>;
-  readonly turns: Array<{ id: TurnId; items: Array<unknown> }>;
-  activeAbort: AbortController | undefined;
-  activeAssistantItems: Map<string, string>;
-  activeCompletedAssistantItems: Set<string>;
-  activeTokenUsage: TokenUsageUpdate | undefined;
-}
-
-export interface DroidAdapterOptions {
-  readonly instanceId?: ProviderInstanceId;
-  readonly environment?: NodeJS.ProcessEnv;
-  readonly sdk?: {
-    readonly createSession: (options?: CreateSessionOptions) => Promise<DroidSession>;
-    readonly resumeSession: (
-      sessionId: string,
-      options?: ResumeSessionOptions,
-    ) => Promise<DroidSession>;
-  };
-}
-
-const nowIso = () => DateTime.formatIso(DateTime.nowUnsafe());
-const eventId = () => EventId.make(randomUUID());
-const SUPPORTED_DROID_IMAGE_MIME_TYPES = [
-  "image/gif",
-  "image/jpeg",
-  "image/png",
-  "image/webp",
-] as const;
-type SupportedDroidImageMimeType = (typeof SUPPORTED_DROID_IMAGE_MIME_TYPES)[number];
-const isSupportedDroidImageMimeType = (value: string): value is SupportedDroidImageMimeType =>
-  (SUPPORTED_DROID_IMAGE_MIME_TYPES as ReadonlyArray<string>).includes(value);
-
-function updateContextSession(context: DroidContext, patch: Partial<ProviderSession>) {
-  context.session = {
-    ...context.session,
-    ...patch,
-    updatedAt: nowIso(),
-  };
-}
-
-function toModelId(model: string | undefined): string | undefined {
-  return !model || model === "default" ? undefined : model;
-}
-
-function toReasoningEffort(value: string | undefined): ReasoningEffort | undefined {
-  switch (value) {
-    case "none":
-      return ReasoningEffort.None;
-    case "dynamic":
-      return ReasoningEffort.Dynamic;
-    case "off":
-      return ReasoningEffort.Off;
-    case "minimal":
-      return ReasoningEffort.Minimal;
-    case "low":
-      return ReasoningEffort.Low;
-    case "medium":
-      return ReasoningEffort.Medium;
-    case "high":
-      return ReasoningEffort.High;
-    case "xhigh":
-      return ReasoningEffort.ExtraHigh;
-    case "max":
-      return ReasoningEffort.Max;
-    default:
-      return undefined;
-  }
-}
-
-function toAutonomyLevel(input: ProviderSessionStartInput): AutonomyLevel {
-  switch (input.runtimeMode) {
-    case "approval-required":
-      return AutonomyLevel.Off;
-    case "auto-accept-edits":
-      return AutonomyLevel.Low;
-    case "medium-access":
-      return AutonomyLevel.Medium;
-    case "full-access":
-      return AutonomyLevel.High;
-  }
-}
-
-function contentBlockText(block: ContentBlock): string {
-  if (block.type === "text") return block.text;
-  if (block.type === "thinking") return block.thinking;
-  return "";
-}
-
-function toRequestType(params: RequestPermissionRequestParams): CanonicalRequestType {
-  const type = params.toolUses[0]?.confirmationType;
-  switch (type) {
-    case ToolConfirmationType.Execute:
-      return "command_execution_approval";
-    case ToolConfirmationType.Edit:
-    case ToolConfirmationType.Create:
-    case ToolConfirmationType.ApplyPatch:
-      return "file_change_approval";
-    case ToolConfirmationType.McpTool:
-      return "dynamic_tool_call";
-    case ToolConfirmationType.AskUser:
-      return "tool_user_input";
-    default:
-      return "unknown";
-  }
-}
-
-function toToolItemType(toolName: string): ToolLifecycleItemType {
-  const normalized = toolName.toLowerCase();
-  if (
-    normalized.includes("exec") ||
-    normalized.includes("bash") ||
-    normalized.includes("command")
-  ) {
-    return "command_execution";
-  }
-  if (normalized.includes("edit") || normalized.includes("write") || normalized.includes("patch")) {
-    return "file_change";
-  }
-  if (normalized.includes("mcp")) return "mcp_tool_call";
-  if (normalized.includes("web")) return "web_search";
-  if (normalized.includes("image")) return "image_view";
-  return "dynamic_tool_call";
-}
-
-function permissionDetail(params: RequestPermissionRequestParams): string {
-  const first = params.toolUses[0];
-  if (!first) return "Droid requested permission.";
-  const details = first.details;
-  switch (details.type) {
-    case ToolConfirmationType.Execute:
-      return details.fullCommand;
-    case ToolConfirmationType.Edit:
-    case ToolConfirmationType.Create:
-    case ToolConfirmationType.ApplyPatch:
-      return "filePath" in details ? details.filePath : "Droid requested a file change.";
-    case ToolConfirmationType.McpTool:
-      return details.toolName;
-    default:
-      return first.toolUse.name;
-  }
-}
-
-function normalizeAskUserQuestions(params: AskUserRequestParams): ReadonlyArray<UserInputQuestion> {
-  return params.questions.map((question, index) => ({
-    id: `question-${question.index ?? index}`,
-    header: question.topic || `Question ${index + 1}`,
-    question: question.question,
-    options: question.options.map((option) => ({
-      label: option,
-      description: option,
-    })),
-  }));
-}
-
-function answerString(value: unknown): string {
-  if (Array.isArray(value)) return value.map(answerString).join(", ");
-  return typeof value === "string" ? value : value == null ? "" : JSON.stringify(value);
-}
-
-function toAskUserResult(
-  questions: AskUserRequestParams["questions"],
-  answers: ProviderUserInputAnswers,
-): AskUserResult {
-  return {
-    answers: questions.map((question, index) => ({
-      index: question.index,
-      question: question.question,
-      answer: answerString(
-        answers[`question-${question.index ?? index}`] ?? answers[question.question],
-      ),
-    })),
-  };
-}
-
-function toOutcome(decision: ProviderApprovalDecision): ToolConfirmationOutcome {
-  switch (decision) {
-    case "accept":
-      return ToolConfirmationOutcome.ProceedOnce;
-    case "acceptForSession":
-      return ToolConfirmationOutcome.ProceedAlways;
-    case "decline":
-    case "cancel":
-      return ToolConfirmationOutcome.Cancel;
-  }
-}
-
-function toTokenUsageSnapshot(usage: TokenUsageUpdate) {
-  const inputTokens = usage.inputTokens + usage.cacheCreationTokens + usage.cacheReadTokens;
-  const outputTokens = usage.outputTokens + usage.thinkingTokens;
-  return {
-    usedTokens: inputTokens + outputTokens,
-    inputTokens,
-    cachedInputTokens: usage.cacheReadTokens,
-    outputTokens,
-    reasoningOutputTokens: usage.thinkingTokens,
-    lastInputTokens: inputTokens,
-    lastCachedInputTokens: usage.cacheReadTokens,
-    lastOutputTokens: outputTokens,
-    lastReasoningOutputTokens: usage.thinkingTokens,
-  };
-}
+export type { DroidAdapterOptions } from "../droid/DroidAdapterTypes.ts";
 
 export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapterOptions) {
   return Effect.gen(function* () {
@@ -312,83 +92,16 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
     const emit = (event: ProviderRuntimeEvent) =>
       Queue.offer(runtimeEvents, event).pipe(Effect.asVoid);
     const emitNow = (event: ProviderRuntimeEvent) => runPromise(emit(event));
-    const eventBase = (
-      context: DroidContext,
-      input?: {
-        turnId?: TurnId;
-        itemId?: string;
-        requestId?: string;
-        raw?: unknown;
-      },
-    ) => ({
-      eventId: eventId(),
-      provider: PROVIDER,
-      providerInstanceId: instanceId,
-      threadId: context.session.threadId,
-      createdAt: nowIso(),
-      ...(input?.turnId ? { turnId: input.turnId } : {}),
-      ...(input?.itemId ? { itemId: RuntimeItemId.make(input.itemId) } : {}),
-      ...(input?.requestId ? { requestId: RuntimeRequestId.make(input.requestId) } : {}),
-      ...(input?.raw !== undefined
-        ? { raw: { source: "droid.sdk.message" as const, payload: input.raw } }
-        : {}),
-    });
+    const eventBase = makeDroidEventBase(instanceId);
     const requireSession = Effect.fn("requireDroidSession")(function* (threadId: ThreadId) {
       const context = sessions.get(threadId);
       if (!context) {
         return yield* new ProviderAdapterSessionNotFoundError({
-          provider: PROVIDER,
+          provider: DROID_PROVIDER,
           threadId,
         });
       }
       return context;
-    });
-
-    type DroidAdapterShape = ProviderAdapterShape<ProviderAdapterError>;
-    const resolveImages = Effect.fn("resolveDroidImages")(function* (
-      input: NonNullable<Parameters<DroidAdapterShape["sendTurn"]>[0]["attachments"]>,
-    ) {
-      return yield* Effect.forEach(
-        input,
-        (attachment) =>
-          Effect.gen(function* () {
-            if (!isSupportedDroidImageMimeType(attachment.mimeType)) {
-              return yield* new ProviderAdapterRequestError({
-                provider: PROVIDER,
-                method: "turn/start",
-                detail: `Unsupported Droid image attachment type '${attachment.mimeType}'.`,
-              });
-            }
-            const attachmentPath = resolveAttachmentPath({
-              attachmentsDir: serverConfig.attachmentsDir,
-              attachment,
-            });
-            if (!attachmentPath) {
-              return yield* new ProviderAdapterRequestError({
-                provider: PROVIDER,
-                method: "turn/start",
-                detail: `Invalid attachment id '${attachment.id}'.`,
-              });
-            }
-            const bytes = yield* fileSystem.readFile(attachmentPath).pipe(
-              Effect.mapError(
-                (cause) =>
-                  new ProviderAdapterRequestError({
-                    provider: PROVIDER,
-                    method: "turn/start",
-                    detail: `Failed to read attachment file: ${cause.message}.`,
-                    cause,
-                  }),
-              ),
-            );
-            return {
-              type: "base64",
-              data: Buffer.from(bytes).toString("base64"),
-              mediaType: attachment.mimeType,
-            } satisfies Base64ImageSource;
-          }),
-        { concurrency: 1 },
-      );
     });
 
     const startSession: DroidAdapterShape["startSession"] = Effect.fn("startDroidSession")(
@@ -461,14 +174,14 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
                 }),
           catch: (cause) =>
             new ProviderAdapterRequestError({
-              provider: PROVIDER,
+              provider: DROID_PROVIDER,
               method: "createSession",
               detail: cause instanceof Error ? cause.message : "Failed to start Droid session.",
               cause,
             }),
         });
         const session: ProviderSession = {
-          provider: PROVIDER,
+          provider: DROID_PROVIDER,
           providerInstanceId: instanceId,
           status: "ready",
           runtimeMode: input.runtimeMode,
@@ -507,200 +220,23 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
       },
     );
 
-    const handleMessage = async (context: DroidContext, turnId: TurnId, message: DroidMessage) => {
-      const base = (itemId?: string) =>
-        eventBase(context, { turnId, raw: message, ...(itemId ? { itemId } : {}) });
-      switch (message.type) {
-        case DroidMessageType.AssistantTextDelta:
-        case DroidMessageType.ThinkingTextDelta: {
-          const itemId = `${message.messageId}-${message.blockIndex}`;
-          const streamKind: RuntimeContentStreamKind =
-            message.type === DroidMessageType.AssistantTextDelta
-              ? "assistant_text"
-              : "reasoning_text";
-          if (streamKind === "assistant_text") {
-            context.activeAssistantItems.set(
-              itemId,
-              `${context.activeAssistantItems.get(itemId) ?? ""}${message.text}`,
-            );
-          }
-          return emitNow({
-            ...base(itemId),
-            type: "content.delta",
-            payload: { streamKind, delta: message.text },
-          });
-        }
-        case DroidMessageType.CreateMessage: {
-          if (message.role !== "assistant") {
-            return;
-          }
-          for (const [index, block] of message.content.entries()) {
-            const text = contentBlockText(block);
-            if (text.length === 0) {
-              continue;
-            }
-            const itemId = block.id ?? `${message.messageId}-${index}`;
-            if (block.type === "text") {
-              const previousText = context.activeAssistantItems.get(itemId) ?? "";
-              const delta = text.startsWith(previousText) ? text.slice(previousText.length) : text;
-              if (delta.length > 0) {
-                await emitNow({
-                  ...base(itemId),
-                  type: "content.delta",
-                  payload: { streamKind: "assistant_text", delta },
-                });
-              }
-              context.activeAssistantItems.set(itemId, text);
-              continue;
-            }
-            if (block.type === "thinking") {
-              await emitNow({
-                ...base(itemId),
-                type: "content.delta",
-                payload: { streamKind: "reasoning_text", delta: text },
-              });
-            }
-          }
-
-          const firstTextIndex = message.content.findIndex((block) => block.type === "text");
-          const firstTextBlock = message.content[firstTextIndex];
-          const completedItemId =
-            firstTextBlock?.id ??
-            (firstTextIndex >= 0 ? `${message.messageId}-${firstTextIndex}` : message.messageId);
-          if (!context.activeCompletedAssistantItems.has(completedItemId)) {
-            context.activeCompletedAssistantItems.add(completedItemId);
-            return emitNow({
-              ...base(completedItemId),
-              type: "item.completed",
-              payload: {
-                itemType: "assistant_message",
-                status: "completed",
-                ...(firstTextBlock ? { detail: contentBlockText(firstTextBlock) } : {}),
-              },
-            });
-          }
-          return;
-        }
-        case DroidMessageType.ToolUse:
-          return emitNow({
-            ...base(message.toolUseId),
-            type: "item.started",
-            payload: {
-              itemType: toToolItemType(message.toolName),
-              status: "inProgress",
-              title: message.toolName,
-              data: message.toolInput,
-            },
-          });
-        case DroidMessageType.ToolProgress:
-          return emitNow({
-            ...base(message.toolUseId),
-            type: "item.updated",
-            payload: {
-              itemType: toToolItemType(message.toolName),
-              status: "inProgress",
-              title: message.toolName,
-              detail: message.content,
-              data: message.update,
-            },
-          });
-        case DroidMessageType.ToolResult:
-          return emitNow({
-            ...base(message.toolUseId),
-            type: "item.completed",
-            payload: {
-              itemType: toToolItemType(message.toolName),
-              status: message.isError ? "failed" : "completed",
-              title: message.toolName,
-              detail:
-                typeof message.content === "string"
-                  ? message.content
-                  : JSON.stringify(message.content),
-            },
-          });
-        case DroidMessageType.WorkingStateChanged:
-          return emitNow({
-            ...base(),
-            type: "session.state.changed",
-            payload: {
-              state:
-                message.state === "idle"
-                  ? "ready"
-                  : message.state.includes("waiting")
-                    ? "waiting"
-                    : "running",
-              detail: message,
-            },
-          });
-        case DroidMessageType.TokenUsageUpdate:
-          context.activeTokenUsage = message;
-          return emitNow({
-            ...base(),
-            type: "thread.token-usage.updated",
-            payload: { usage: toTokenUsageSnapshot(message) },
-          });
-        case DroidMessageType.SessionTitleUpdated:
-          return emitNow({
-            ...base(),
-            type: "thread.metadata.updated",
-            payload: { name: message.title },
-          });
-        case DroidMessageType.SettingsUpdated:
-          return emitNow({
-            ...base(),
-            type: "session.configured",
-            payload: { config: message.settings },
-          });
-        case DroidMessageType.McpStatusChanged:
-          return emitNow({
-            ...base(),
-            type: "mcp.status.updated",
-            payload: { status: message },
-          });
-        case DroidMessageType.McpAuthRequired:
-          return emitNow({
-            ...base(),
-            type: "auth.status",
-            payload: { isAuthenticating: true, output: [message.message] },
-          });
-        case DroidMessageType.McpAuthCompleted:
-          return emitNow({
-            ...base(),
-            type: "mcp.oauth.completed",
-            payload: {
-              success: message.outcome === "success",
-              name: message.serverName,
-              ...(message.outcome === "success" ? {} : { error: message.message }),
-            },
-          });
-        case DroidMessageType.Error:
-          return emitNow({
-            ...base(),
-            type: "runtime.error",
-            payload: { message: message.message, class: "provider_error" },
-          });
-        case DroidMessageType.TurnComplete:
-          if (message.tokenUsage) context.activeTokenUsage = message.tokenUsage;
-          return Promise.resolve();
-        default:
-          return Promise.resolve();
-      }
-    };
-
     const sendTurn: DroidAdapterShape["sendTurn"] = Effect.fn("sendDroidTurn")(function* (input) {
       const context = sessions.get(input.threadId);
       if (!context) {
         return yield* new ProviderAdapterValidationError({
-          provider: PROVIDER,
+          provider: DROID_PROVIDER,
           operation: "sendTurn",
           issue: `Unknown Droid thread: ${input.threadId}`,
         });
       }
       const text = input.input?.trim();
-      const images = yield* resolveImages(input.attachments ?? []);
+      const images = yield* resolveDroidImages(input.attachments ?? [], {
+        attachmentsDir: serverConfig.attachmentsDir,
+        fileSystem,
+      });
       if (!text && images.length === 0) {
         return yield* new ProviderAdapterValidationError({
-          provider: PROVIDER,
+          provider: DROID_PROVIDER,
           operation: "sendTurn",
           issue: "Droid turns require text input or at least one attachment.",
         });
@@ -713,7 +249,7 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
       context.activeCompletedAssistantItems = new Set();
       context.activeTokenUsage = undefined;
       context.turns.push({ id: turnId, items: [] });
-      updateContextSession(context, {
+      updateDroidContextSession(context, {
         status: "running",
         activeTurnId: turnId,
         model: input.modelSelection?.model ?? context.session.model,
@@ -755,7 +291,7 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
             text || "Please respond to the attached image.",
             messageOptions,
           )) {
-            await handleMessage(context, turnId, message);
+            await handleDroidMessage({ context, turnId, message, eventBase, emitNow });
           }
           for (const [itemId, detail] of context.activeAssistantItems) {
             if (context.activeCompletedAssistantItems.has(itemId)) {
@@ -767,7 +303,7 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
               payload: { itemType: "assistant_message", status: "completed", detail },
             });
           }
-          updateContextSession(context, { status: "ready", activeTurnId: undefined });
+          updateDroidContextSession(context, { status: "ready", activeTurnId: undefined });
           await emitNow({
             ...eventBase(context, { turnId }),
             type: "turn.completed",
@@ -778,7 +314,7 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
           });
         } catch (cause) {
           const message = cause instanceof Error ? cause.message : "Droid turn failed.";
-          updateContextSession(context, {
+          updateDroidContextSession(context, {
             status: "error",
             activeTurnId: undefined,
             lastError: message,
@@ -814,7 +350,7 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
       });
 
     return {
-      provider: PROVIDER,
+      provider: DROID_PROVIDER,
       capabilities: { sessionModelSwitch: "in-session" },
       startSession,
       sendTurn,
@@ -830,7 +366,7 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
           const pending = context?.pendingPermissions.get(requestId);
           if (!context || !pending) {
             return yield* new ProviderAdapterRequestError({
-              provider: PROVIDER,
+              provider: DROID_PROVIDER,
               method: "respondToRequest",
               detail: `Unknown pending Droid permission request: ${requestId}`,
             });
@@ -849,7 +385,7 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
           const pending = context?.pendingUserInputs.get(requestId);
           if (!context || !pending) {
             return yield* new ProviderAdapterRequestError({
-              provider: PROVIDER,
+              provider: DROID_PROVIDER,
               method: "respondToUserInput",
               detail: `Unknown pending Droid user-input request: ${requestId}`,
             });
@@ -875,7 +411,7 @@ export function makeDroidAdapter(settings: DroidSettings, options?: DroidAdapter
           const context = yield* requireSession(threadId);
           if (!Number.isInteger(numTurns) || numTurns < 1) {
             return yield* new ProviderAdapterValidationError({
-              provider: PROVIDER,
+              provider: DROID_PROVIDER,
               operation: "rollbackThread",
               issue: "numTurns must be an integer >= 1.",
             });
